# generated/api/model.go
// Package api - Auto-generated models
// DO NOT EDIT - Generated by restful-api-gen-go
package api

import "time"

// User represents the users table
type User struct {
	// Primary key
	Id int64 `json:"id" db:"id"`
	// Full name
	Name string `json:"name" db:"name"`
	// Email address
	Email string `json:"email" db:"email"`
	// bcrypt hash
	PasswordHash string `json:"password_hash" db:"password_hash"`
	// active, inactive, suspended
	Status string `json:"status" db:"status"`
	// Record created
	CreatedAt time.Time `json:"created_at" db:"created_at"`
	// Record updated
	UpdatedAt *time.Time `json:"updated_at,omitempty" db:"updated_at"`
}

func (User) TableName() string { return "users" }

// UserRelations holds related records
type UserRelations struct {
	Order []Order `json:"orders,omitempty"`
}

// Order represents the orders table
type Order struct {
	// id
	Id int64 `json:"id" db:"id"`
	// FK to users
	UserId int64 `json:"user_id" db:"user_id"`
	// Order total
	TotalAmount float64 `json:"total_amount" db:"total_amount"`
	// status
	Status string `json:"status" db:"status"`
	// created_at
	CreatedAt time.Time `json:"created_at" db:"created_at"`
	// updated_at
	UpdatedAt *time.Time `json:"updated_at,omitempty" db:"updated_at"`
}

func (Order) TableName() string { return "orders" }