# generated/openapi.yaml
openapi: "3.0.3"
info:
  title: "github.com/yourorg/your-api"
  description: "Auto-generated RESTful API from table.json"
  version: "v1"
servers:
  - url: http://localhost:8080/api/v1
    description: "Local development"

paths:
  /users:
    get:
      operationId: listUser
      summary: "List all users records"
      tags:
        - users
      parameters:
        - name: page
          in: query
          description: "Page number"
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          description: "Items per page (max 100)"
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: "Paginated list of users"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
                  total_pages:
                    type: integer

    post:
      operationId: createUser
      summary: "Create a new users record"
      tags:
        - users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "201":
          description: "Created User"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: "Bad request"

  /users/{id}:
    get:
      operationId: getUser
      summary: "Get a single User by ID"
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          description: "User ID"
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "The requested User"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: "User not found"

    put:
      operationId: updateUser
      summary: "Update an existing User record"
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          description: "User ID"
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: "Updated User"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: "Bad request"
        "404":
          description: "User not found"

    delete:
      operationId: deleteUser
      summary: "Delete a User record"
      tags:
        - users
      parameters:
        - name: id
          in: path
          required: true
          description: "User ID"
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: "User deleted successfully"
        "404":
          description: "User not found"

  /orders:
    get:
      operationId: listOrder
      summary: "List all orders records"
      tags:
        - orders
      parameters:
        - name: page
          in: query
          description: "Page number"
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          description: "Items per page (max 100)"
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: "Paginated list of orders"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Order"
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
                  total_pages:
                    type: integer

    post:
      operationId: createOrder
      summary: "Create a new orders record"
      tags:
        - orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Order"
      responses:
        "201":
          description: "Created Order"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: "Bad request"

  /orders/{id}:
    get:
      operationId: getOrder
      summary: "Get a single Order by ID"
      tags:
        - orders
      parameters:
        - name: id
          in: path
          required: true
          description: "Order ID"
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "The requested Order"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: "Order not found"

    put:
      operationId: updateOrder
      summary: "Update an existing Order record"
      tags:
        - orders
      parameters:
        - name: id
          in: path
          required: true
          description: "Order ID"
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Order"
      responses:
        "200":
          description: "Updated Order"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: "Bad request"
        "404":
          description: "Order not found"

    delete:
      operationId: deleteOrder
      summary: "Delete a Order record"
      tags:
        - orders
      parameters:
        - name: id
          in: path
          required: true
          description: "Order ID"
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: "Order deleted successfully"
        "404":
          description: "Order not found"

components:
  schemas:
    User:
      type: object
      description: "Represents the users table"
      properties:
        id:
          type: integer
          format: int64
          description: "Primary key"
          readOnly: true
        name:
          type: string
          description: "Full name"
          maxLength: 100
        email:
          type: string
          description: "Email address"
          maxLength: 255
        password_hash:
          type: string
          description: "bcrypt hash"
          maxLength: 255
        status:
          type: string
          description: "active, inactive, suspended"
          maxLength: 20
        created_at:
          type: string
          format: date-time
          description: "Record created"
        updated_at:
          type: string
          format: date-time
          nullable: true
          description: "Record updated"

    Order:
      type: object
      description: "Represents the orders table"
      properties:
        id:
          type: integer
          format: int64
          description: "id"
          readOnly: true
        user_id:
          type: integer
          format: int64
          description: "FK to users"
        total_amount:
          type: number
          format: double
          description: "Order total"
        status:
          type: string
          description: "status"
          maxLength: 20
        created_at:
          type: string
          format: date-time
          description: "created_at"
        updated_at:
          type: string
          format: date-time
          nullable: true
          description: "updated_at"